refactor api key tests
